Match func_800569AC .rodata by splitting 4FEB0 at 57F20
