+++ versions/us1/mischiefmakers.yaml
@@ -109,6 +109,7 @@ segments:
       - [0x048A30, c]
       - [0x049200, c, marina]
       - [0x04FEB0, c]
+      - [0x057F20, c]
       - [0x059EA0, asm]
       - [0x05D120, asm]
       - [0x05E230, asm]
@@ -372,6 +373,7 @@ segments:
       - [0x0EC920, rodata] # 438E0?
       - [0x0EC960, rodata] # marina?
       - [0x0EC9D0, .rodata, 4FEB0]
+      - [0x0ECAA0, .rodata, 57F20]
       - [0x0ECAC0, rodata] # 59EA0?
       - [0x0ECB90, rodata] # 5E230?
       - [0x0ECCF0, rodata] # 60CF0?
